Properly filter out stale interfaces
I believe this is a partial fix for the following issues:
https://github.com/elm-lang/elm-compiler/issues/844
https://github.com/elm-lang/elm-make/issues/4

I think there is still a problem when modules have the same name and
are in the same package. This primarily comes up with modules named
`Main` but I think the bigger solution is to mark interface files with
some extra information.