Fix #107, make sure cached interfaces for dependencies are older
If the cached .elmi file is older than the source file, it is invalid.
So we go through and collect all valid .elmi files.

If a cached .elmi file is older than the .elmi files it depends on, it
is also invalid. This was the case that was missing. So now we check if
all the dependencies are transitively fresh, and if so, that they are
all older than our .elmi